+++ Runtime/Framework/Watch/WatchCollection.cs
@@ -8,7 +8,7 @@ namespace DT.UniStart {
   /// <summary>
   /// Watch a list-like type for changes.
   /// </summary>
-  public class WatchIList<L, T> : WatchRef<L>, IList<T>, IWatchable, IWatchable<ReadOnlyCollection<T>>, IGetValue<ReadOnlyCollection<T>> where L : IList<T> {
+  public class WatchIList<L, T> : WatchRef<L>, IList<T>, IReadOnlyList<T>, IWatchable, IWatchable<ReadOnlyCollection<T>>, IGetValue<ReadOnlyCollection<T>> where L : IList<T> {
     LazyComputed<ReadOnlyCollection<T>> readOnlyList;
     AdvancedEvent<ReadOnlyCollection<T>> onChange;
 
@@ -76,7 +76,7 @@ namespace DT.UniStart {
   /// <summary>
   /// Watch a dictionary-like type for changes.
   /// </summary>
-  public class WatchIDictionary<D, K, V> : WatchRef<D>, IDictionary<K, V>, IWatchable, IWatchable<ReadOnlyDictionary<K, V>>, IGetValue<ReadOnlyDictionary<K, V>> where D : IDictionary<K, V> {
+  public class WatchIDictionary<D, K, V> : WatchRef<D>, IDictionary<K, V>, IReadOnlyDictionary<K, V>, IWatchable, IWatchable<ReadOnlyDictionary<K, V>>, IGetValue<ReadOnlyDictionary<K, V>>, IDictionaryState<K, V> where D : IDictionary<K, V> {
     LazyComputed<ReadOnlyDictionary<K, V>> readOnlyDictionary;
     AdvancedEvent<ReadOnlyDictionary<K, V>> onChange;
 
@@ -140,6 +140,8 @@ namespace DT.UniStart {
     IEnumerator IEnumerable.GetEnumerator() => this.value.GetEnumerator();
     public ICollection<K> Keys => this.value.Keys;
     public ICollection<V> Values => this.value.Values;
+    IEnumerable<K> IReadOnlyDictionary<K, V>.Keys { get; }
+    IEnumerable<V> IReadOnlyDictionary<K, V>.Values { get; }
     public bool IsReadOnly => this.value.IsReadOnly;
     #endregion
   }
@@ -148,7 +150,7 @@ namespace DT.UniStart {
   /// Watch a list for changes.
   /// </summary>
   [Serializable]
-  public class WatchList<T> : WatchIList<List<T>, T>, IWatchable, IWatchable<ReadOnlyCollection<T>> {
+  public class WatchList<T> : WatchIList<List<T>, T>, IWatchable, IWatchable<ReadOnlyCollection<T>>, IListState<T> {
     public WatchList() : base(new List<T>()) { }
     public WatchList(List<T> value) : base(value) { }
 
@@ -162,7 +164,7 @@ namespace DT.UniStart {
   /// Watch an array for changes.
   /// </summary>
   [Serializable]
-  public class WatchArray<T> : WatchIList<T[], T>, IWatchable, IWatchable<ReadOnlyCollection<T>> {
+  public class WatchArray<T> : WatchIList<T[], T>, IWatchable, IWatchable<ReadOnlyCollection<T>>, IArrayState<T> {
     public WatchArray(int n) : base(new T[n]) { }
     public WatchArray(T[] value) : base(value) { }
 
@@ -175,7 +177,7 @@ namespace DT.UniStart {
   /// <summary>
   /// Watch a dictionary for changes.
   /// </summary>
-  public class WatchDictionary<K, V> : WatchIDictionary<Dictionary<K, V>, K, V>, IWatchable, IWatchable<ReadOnlyDictionary<K, V>> {
+  public class WatchDictionary<K, V> : WatchIDictionary<Dictionary<K, V>, K, V>, IWatchable, IWatchable<ReadOnlyDictionary<K, V>>, IDictionaryState<K, V> {
     public WatchDictionary() : base(new Dictionary<K, V>()) { }
     public WatchDictionary(Dictionary<K, V> value) : base(value) { }
   }